+++ superset-frontend/src/dashboard/components/dnd/DragDroppable.jsx
@@ -37,7 +37,7 @@ const propTypes = {
   component: componentShape.isRequired,
   parentComponent: componentShape,
   depth: PropTypes.number.isRequired,
-  disableDragDrop: PropTypes.bool,
+  disableDragdrop: PropTypes.bool,
   orientation: PropTypes.oneOf(['row', 'column']),
   index: PropTypes.number.isRequired,
   style: PropTypes.object,
@@ -58,7 +58,7 @@ const defaultProps = {
   className: null,
   style: null,
   parentComponent: null,
-  disableDragDrop: false,
+  disableDragdrop: false,
   children() {},
   onDrop() {},
   orientation: 'row',
