+++ superset-frontend/src/dashboard/components/nativeFilters/FilterBar/FilterControls/FilterValue.tsx
@@ -30,7 +30,7 @@ import { getChartDataRequest } from 'src/chart/chartAction';
 import Loading from 'src/components/Loading';
 import BasicErrorAlert from 'src/components/ErrorMessage/BasicErrorAlert';
 import { FilterProps } from './types';
-import { getFormData } from '../utils';
+import { getFormData } from '../../utils';
 import { useCascadingFilters } from './state';
 
 const StyledLoadingBox = styled.div`
